[REFACTOR] 모든 styles 파일을 제거
