--- components/Dashboard/Graph.styles.ts
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 46px;
-
-  width: 1200px;
-`;
-
-export { Container };
