+++ components/Dashboard/GraphSide.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import * as S from './GraphSide.styles';
+
+import styled from '@emotion/styled';
 
 const GraphSide = () => {
   return (
@@ -46,3 +47,81 @@ const GraphSide = () => {
 };
 
 export default GraphSide;
+
+const S = {
+  Container: styled.div`
+    display: flex;
+    flex-direction: column;
+
+    width: 151px;
+  `,
+
+  SerchedDataBox: styled.div`
+    display: flex;
+    flex-direction: column;
+
+    width: 151px;
+    height: 128px;
+  `,
+
+  SerchedDatas: styled.ul`
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+  `,
+
+  SerchedData: styled.li`
+    display: flex;
+    flex-direction: row;
+
+    justify-content: space-between;
+  `,
+
+  DeviceDataBox: styled.div`
+    display: flex;
+    flex-direction: column;
+
+    width: 151px;
+    height: 81px;
+  `,
+
+  DeviceDatas: styled.ul`
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+  `,
+
+  DeviceData: styled.li`
+    display: flex;
+    flex-direction: row;
+
+    justify-content: space-between;
+  `,
+
+  Title: styled.strong`
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 16px;
+    line-height: 19px;
+    /* identical to box height */
+
+    /* Black */
+
+    color: #181818;
+  `,
+
+  Line: styled.div`
+    width: 151px;
+    height: 0px;
+
+    margin-top: 8.5px;
+    margin-bottom: 10.5px;
+
+    border: 1px solid #e9e9e9;
+  `,
+
+  Layout: styled.div`
+    height: 40px;
+  `,
+};
