--- components/Project/install/CodeParagraph.styles.ts
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-const InstallGuide = styled.p`
-  margin-bottom: 30px;
-  font-weight: 500;
-  font-size: 18px;
-`;
-const CodeParagraphWrap = styled.div`
-  margin-bottom: 92px;
-`;
-
-export { InstallGuide, CodeParagraphWrap };
