+++ components/Project/install/CodeParagraph.tsx
@@ -1,7 +1,8 @@
 import { ReactNode } from 'react';
 
 import CodeBlock from 'components/common/CodeBlock/CodeBlock';
-import * as S from './CodeParagraph.styles';
+
+import styled from '@emotion/styled';
 
 interface CodeParagraphProps {
   text: ReactNode;
@@ -18,3 +19,14 @@ const CodeParagraph = ({ text, code }: CodeParagraphProps) => {
 };
 
 export default CodeParagraph;
+
+const S = {
+  InstallGuide: styled.p`
+    margin-bottom: 30px;
+    font-weight: 500;
+    font-size: 18px;
+  `,
+  CodeParagraphWrap: styled.div`
+    margin-bottom: 92px;
+  `,
+};
