--- components/Project/install/InstallPageButtons.styles.ts
@@ -1,8 +1,0 @@
-import styled from '@emotion/styled';
-
-const ButtonsWrap = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-export { ButtonsWrap };
