+++ components/Project/install/common/Button.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
 import { Colors } from 'styles/theme';
-import * as S from './Button.styles';
+import styled from '@emotion/styled';
 
 export type CommonButtonStyleAttributes = {
   width: number;
@@ -29,3 +29,24 @@ const CommonButton = ({
 };
 
 export default CommonButton;
+
+const S = {
+  Button: styled.button<CommonButtonStyleAttributes>`
+    border: none;
+    border-radius: 15px;
+    font-weight: 500;
+    font-size: 18px;
+    overflow: hidden;
+
+    :hover {
+      cursor: pointer;
+      opacity: 0.8;
+    }
+
+    width: ${({ width }) => width}px;
+    height: ${({ height }) => height}px;
+    color: ${({ theme, color }) => theme.colors[color]};
+    background-color: ${({ theme, backgroundColor }) =>
+      theme.colors[backgroundColor]};
+  `,
+};
