+++ components/ShoppingMallReview/Review.tsx
@@ -1,22 +1,16 @@
 import Image from 'next/image';
+import { useRouter } from 'next/router';
+import { useState } from 'react';
+import styled from '@emotion/styled';
 
 import HeartFull from 'public/icons/like-heart-full.svg';
 import HeartEmpty from 'public/icons/like-heart-empty.svg';
-
-import grayStar from 'public/images/empty_star_img.png';
 import thumbsUp from 'public/icons/thumbs_up.png';
-import * as S from './Review.styles';
+
 import { ShoppingMallReviewDetail } from 'typings/reviews';
-import {
-  formattedCreateDate,
-  formattedCreateDateArr,
-  formattedImageUrl,
-  formattedLastTime,
-} from 'utils/format';
+import { formattedCreateDateArr, formattedImageUrl } from 'utils/format';
 import StarBox from 'components/social/common/Review/StarBox';
 import Card from 'components/common/Card';
-import { Router, useRouter } from 'next/router';
-import { useState } from 'react';
 import { shoppingAPI } from 'api/reviews';
 
 interface ReviewProps {
@@ -130,3 +124,200 @@ const Review = ({ reviewDetail }: ReviewProps) => {
 };
 
 export default Review;
+
+const S = {
+  Container: styled.div`
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 30px;
+  `,
+
+  WriterInfo: styled.div`
+    display: flex;
+    flex-direction: row;
+
+    gap: 8px;
+  `,
+
+  ProfileImage: styled(Image)`
+    position: relative;
+    width: 50px;
+    height: 50px;
+
+    background: #d9d9d9;
+    border-radius: 50%;
+  `,
+
+  WriterDesc: styled.div`
+    display: flex;
+    flex-direction: column;
+
+    color: #000000;
+    font-style: normal;
+    font-weight: 500;
+  `,
+
+  WriterName: styled.p`
+    font-family: 'Inter';
+    font-size: 18px;
+    line-height: 22px;
+  `,
+
+  SellerName: styled.p`
+    font-family: 'Pretendard';
+    font-size: 12px;
+    line-height: 14px;
+  `,
+
+  StarRateWithDate: styled.div`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-top: 4px;
+    margin-bottom: 4px;
+    gap: 8px;
+  `,
+
+  WriteDate: styled.p`
+    font-family: 'Inter';
+    font-style: normal;
+    font-weight: 500;
+    font-size: 10px;
+    line-height: 12px;
+
+    color: #828282;
+  `,
+
+  Stars: styled.ul`
+    display: flex;
+    flex-direction: row;
+
+    margin-right: 8px;
+
+    li {
+      img {
+        width: 18px;
+        height: 18px;
+      }
+    }
+  `,
+
+  ProductImageBox: styled.div`
+    display: flex;
+    flex-direction: column;
+    margin-top: 24px;
+    p {
+      font-family: 'Pretendard';
+      font-style: normal;
+      font-weight: 500;
+      font-size: 10px;
+      line-height: 12px;
+
+      color: #828282;
+    }
+  `,
+
+  ProductImages: styled.div`
+    width: 424px;
+    height: 80px;
+
+    margin-top: 8px;
+    margin-bottom: 16px;
+
+    display: flex;
+    flex-direction: row;
+    gap: 8px;
+
+    /* Image {
+    width: 100px;
+    height: 80px;
+
+    background: #d9d9d9;
+  } */
+  `,
+
+  ProductDesc: styled.div`
+    width: 413px;
+    max-height: 126px;
+
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-size: 12px;
+    line-height: 120%;
+
+    margin-bottom: 15px;
+
+    color: #333333;
+    strong {
+      font-weight: 700;
+    }
+  `,
+
+  Keywords: styled.div`
+    width: 133px;
+    height: 31px;
+
+    display: flex;
+    flex-direction: column;
+    gap: 7px;
+  `,
+
+  KeywordDetail: styled.div`
+    height: 12px;
+
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-size: 10px;
+    line-height: 12px;
+
+    color: #181818;
+
+    strong {
+      font-weight: 700;
+    }
+    p {
+      font-weight: 400;
+    }
+  `,
+
+  HelpfulRates: styled.div`
+    width: 50%;
+    max-width: 128px;
+
+    margin-top: 21px;
+    margin-bottom: 28px;
+
+    display: flex;
+    flex-direction: row;
+
+    p {
+      font-family: 'Pretendard';
+      font-style: normal;
+      font-weight: 500;
+      font-size: 14px;
+      line-height: 17px;
+
+      color: #333333;
+    }
+
+    button {
+      outline: none;
+      border: none;
+      background-color: transparent;
+
+      img {
+        width: 16px;
+        height: 16px;
+      }
+    }
+  `,
+  LikeButton: styled.button`
+    width: 20px;
+    height: 20px;
+  `,
+};
