--- components/common/Video/Video.styles.ts
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-const Video = styled.video`
-  width: 100%;
-  padding: 0 10%;
-`;
-const Source = styled.source``;
-
-export { Video, Source };
