+++ components/review/ReviewWriteModal/QualitySection/SurveyBox.tsx
@@ -1,4 +1,4 @@
-import * as S from './SurveyBox.styles';
+import styled from '@emotion/styled';
 
 const SurveyBox = () => {
   return (
@@ -38,3 +38,43 @@ const SurveyBox = () => {
 };
 
 export default SurveyBox;
+
+const S = {
+  Box: styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+
+    padding: 12px 0;
+  `,
+
+  Survey: styled.div`
+    display: grid;
+    grid-template-columns: repeat(3, auto);
+
+    width: fit-content;
+    gap: 14px 24px;
+  `,
+
+  SurveyTitle: styled.label`
+    grid-column: 1/4;
+    color: black;
+    padding-right: 6px;
+    font-weight: 700;
+    font-size: 18px;
+
+    white-space: pre;
+  `,
+
+  AnswerBox: styled.div``,
+
+  Answer: styled.div``,
+
+  RadioButton: styled.input`
+    margin-left: 0;
+  `,
+
+  AnswerText: styled.label`
+    margin-left: 4px;
+  `,
+};
