+++ components/review/common/ImageUploadBox.tsx
@@ -1,12 +1,13 @@
 import { useState, useRef, ChangeEvent, MouseEvent, useEffect } from 'react';
 import Image from 'next/image';
+import styled from '@emotion/styled';
 
 import { Colors } from 'styles/theme';
 import { formattedImageUrl } from 'utils/format';
-import * as S from './ImageUploadBox.styles';
 import CameraIcon from 'public/icons/camera.svg';
 import CloseIcon from 'public/icons/close.svg';
 import useHorizontalScroll from 'hooks/useHorizontalScroll';
+import { CommonS as ReviewWriteModalCommonStyles } from 'components/Review/ReviewWriteModal/ReviewWriteModal';
 
 export interface ImageUploadBoxProps {
   buttonColor: Colors;
@@ -124,3 +125,62 @@ const ImageUploadBox = ({
 };
 
 export default ImageUploadBox;
+
+const S = {
+  ImageUploadBox: styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+
+    & > input {
+      display: none;
+    }
+  `,
+
+  ImageUploadButton: styled(ReviewWriteModalCommonStyles.Button)<{
+    color: Colors;
+  }>`
+    display: flex;
+    align-items: center;
+    gap: 10px;
+
+    width: fit-content;
+    font-size: 16px;
+    padding: 7px 15px;
+
+    background: ${({ theme, color }) => theme.colors[color]};
+  `,
+
+  ImageList: styled.div`
+    display: flex;
+    gap: 10px;
+    overflow-x: scroll;
+    width: 790px;
+    height: 85px;
+
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  `,
+
+  ImageBox: styled.div`
+    position: relative;
+    > img {
+      box-shadow: 0 0 5px 1px #d9d6d6;
+    }
+  `,
+
+  CloseButton: styled.button`
+    position: absolute;
+    top: 2px;
+    right: -4px;
+
+    background: none;
+    border: none;
+    outline: none;
+    cursor: pointer;
+  `,
+};
