+++ components/review/common/RatingBox.tsx
@@ -1,6 +1,7 @@
 import Image from 'next/image';
+import styled from '@emotion/styled';
 
-import * as S from './RatingBox.styles';
+import { spinner } from 'styles/animation';
 import StarRating from './StarRating';
 
 interface RatingBoxProps {
@@ -28,7 +29,7 @@ const RatingBox = ({
           <S.Spinner />
         </S.LoadingBox>
       ) : (
-        <S.NullImage>No Image</S.NullImage>
+        <CommonS.NullImage>No Image</CommonS.NullImage>
       )}
       <div>
         <S.ProductName>{productName}</S.ProductName>
@@ -39,3 +40,52 @@ const RatingBox = ({
 };
 
 export default RatingBox;
+
+const CommonS = {
+  NullImage: styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    width: 150px;
+    height: 150px;
+    font-size: 20px;
+    background: ${({ theme }) => theme.colors.gray_1};
+    color: ${({ theme }) => theme.colors.gray_5};
+  `,
+};
+
+const S = {
+  Box: styled.div`
+    display: flex;
+    gap: 20px;
+
+    padding: 12px 0;
+
+    & > div {
+      display: flex;
+      flex-direction: column;
+      gap: 12px;
+
+      font-weight: 500;
+      padding: 5px 0;
+    }
+  `,
+
+  LoadingBox: styled(CommonS.NullImage)``,
+
+  Spinner: styled.div`
+    box-sizing: border-box;
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    border: 4px solid #ccc;
+    border-top-color: ${({ theme }) => theme.colors.blue_0};
+    animation: ${spinner} 1s linear infinite;
+  `,
+
+  ProductName: styled.span`
+    font-size: 16px;
+    color: black;
+  `,
+};
