+++ components/signIn/SocialLoginBox.tsx
@@ -1,4 +1,4 @@
-import * as S from './SocialLoginBox.styles';
+import styled from '@emotion/styled';
 
 import { GoogleOauth } from 'components/common/Oauth/GoogleOauth';
 import { KakaoOauth } from 'components/common/Oauth/KakaoOauth';
@@ -22,3 +22,38 @@ const SocialLoginBox = () => {
 };
 
 export default SocialLoginBox;
+
+const S = {
+  Box: styled.div``,
+
+  TitleBox: styled.div`
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin-bottom: 30px;
+  `,
+
+  Title: styled.span`
+    color: #6d6d6d;
+    background: #f7f7f7;
+    padding: 0 15px;
+    z-index: 1;
+    margin-top: -10px;
+  `,
+
+  DividingLine: styled.div`
+    background: #cacaca;
+    width: 100%;
+    height: 2px;
+  `,
+
+  ButtonBox: styled.div`
+    display: flex;
+    justify-content: center;
+    gap: 40px;
+
+    > * {
+      cursor: pointer;
+    }
+  `,
+};
